+++ src/avt/Expressions/Math/avtSquareExpression.C
@@ -37,7 +37,7 @@
 *****************************************************************************/
 
 // ************************************************************************* //
-//                             avtSquareExpression.C                             //
+//                             avtSquareExpression.C                         //
 // ************************************************************************* //
 
 #include <avtSquareExpression.h>
@@ -107,13 +107,13 @@ avtSquareExpression::~avtSquareExpression()
  
 void
 avtSquareExpression::DoOperation(vtkDataArray *in, vtkDataArray *out,
-                             int ncomponents, int ntuples)
+                                 int ncomponents, int ntuples)
 {
     for (int i = 0 ; i < ntuples ; i++)
     {
         for (int j = 0 ; j < ncomponents ; j++)
         {
-            float val = in->GetComponent(i, j);
+            double val = in->GetComponent(i, j);
             out->SetComponent(i, j, val*val);
         }
     }
