+++ src/plots/Curve/avtOpenGLCurveRenderer.h
@@ -79,8 +79,8 @@ protected:
     void DrawCurveAsSymbols();
     void DrawCurveAsLines();
 
-    void GetAspect(int &bin_x_n, float &bin_x_size, float &bin_x_offset,
-                   int &bin_y_n, float &bin_y_size, float &bin_y_offset);
+    void GetAspect(int &bin_x_n, double &bin_x_size, double &bin_x_offset,
+                   int &bin_y_n, double &bin_y_size, double &bin_y_offset);
     void RenderBall(void);
     void RenderLine(void);
 };
